requirements.txt
# Core API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Data processing & geospatial
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4
geopandas==0.14.0
pyproj==3.6.1
shapely==2.0.2

# Machine learning & interpretability
scikit-learn==1.3.2
xgboost==2.0.3
hdbscan==0.8.34
shap==0.44.1
joblib==1.3.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
geoalchemy2==0.14.1

# Data I/O
pyarrow==14.0.1

# Orchestration
apache-airflow==2.7.3
apache-airflow-providers-postgres==5.10.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
httpx==0.25.2
aiofiles==23.2.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.23.1
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
isort==5.13.2
mypy==1.7.1
pre-commit==3.5.0

# Logging
python-json-logger==2.0.7

# Monitoring
prometheus-client==0.19.0
